correct some indents and stuff
correct some indents and stuff

+++ evince/tools/chocolateyInstall.ps1
@@ -2,5 +2,10 @@ $packageName = '{{PackageName}}'
 $fileType = 'msi'
 $silentArgs = '/quiet'
 $url = '{{DownloadUrl}}' 
-Stop-Process -ProcessName evince*
+#update will fail if evince is on
+$ProcessActive = Get-Process evince* -ErrorAction SilentlyContinue
+if($ProcessActive -ne $null)
+{
+    Stop-Process -ProcessName evince*
+}
 Install-ChocolateyPackage $packageName $fileType $silentArgs $url
